ddi-test/requirements.txt
b2handle==1.1.2
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
ecdsa==0.14.1
humanize==3.2.0
idna==2.10
paramiko==2.7.2
prettytable==2.0.0
psycopg2-binary==2.8.6
pyaml==20.4.0
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.0.1
PyNaCl==1.4.0
python-jose==3.2.0
python-keycloak==0.24.0
PyYAML==5.4.1
requests
robotframework==3.2.2
robotframework-requests==0.8.1
rsa==4.7
six==1.15.0
urllib3==1.26.3
wcwidth==0.2.5
xmlrunner==1.7.7
testarchiver
python-irodsclient@git+https://github.com/lexis-project/python-irodsclient.git@openid_20201105